Fix findContentRelations func to work correctly with the new driver
When cm-neo4j-driver is executing multiple read queries in one go and some of
the queries return cmneo4j.ErrNoResultsFound, all of the queries after that one
won't be executed.
The solution is to either execute each query in a separate Read or to make sure
that it will always return some result. This fix is taking the second approach.